moving string messages to its own dir
